refa(Geral):Create a new Component to manager Legends
- Create Legend component
- Refactoring bar component
